# openapiDoc.JSON
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groupomania social network",
    "description": "This is the documentation of Groupomania social network API.",
    "termsOfService": "http://groupomania.com/terms/",
    "contact": {
      "name": "API support",
      "url": "http://groupomania.com/contact/",
      "email": "contactAPI@groupomania.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [{
    "url": "http://localhost:{port}/{basePath}",
    "variables": {
      "port": {
        "default": "3000"
      },
      "basePath": {
        "default": "api"
      }
    }
  }],
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": ["users"],
        "summary": "Create an account.",
        "description": "Admin creates an account which can be an admin one. Form-data is required if admin put an image file. File type must be 'image/jpeg', 'image/jpg' or 'image/png'.",
        "operationId": "createOneUser",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "User account properties. For more details, go to 'UserPostJson' or 'UserPostFormdata' schema.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userPostJson"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/userPostFormdata"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "get": {
        "tags": ["users"],
        "summary": "Get all users.",
        "description": "Admin gets all user accounts.",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "minDate",
            "in": "query",
            "description": "Get all users who have been created after this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "maxDate",
            "in": "query",
            "description": "Get all users who have been created before this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Get a limited amount of users.",
            "schema": {
              "type": "number",
              "example": 5
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Returns elements from oldest to newest ('asc') or from newest to oldest ('desc').",
            "schema": {
              "type": "string",
              "example": "asc"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "description": "Get users activities.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete all acounts.",
        "description": "Admin deletes all user acounts, except his own.",
        "operationId": "deleteAllAccounts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": ["users"],
        "summary": "Get one account.",
        "description": "User gets properties of his own account, or admin gets properties of a user account.",
        "operationId": "getOneUser",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "userId",
            "in": "path",
            "description": "Id of user to return.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            },
            "required": true
          },
          {
            "name": "activity",
            "in": "query",
            "description": "Get user activity.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request."
          },
          "404": {
            "description": "Invalid Id."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Modify one account.",
        "description": "User modifies properties of his own account, or admin modifies properties of a user account. Form-data is required if an image file is added. File type must be 'image/jpeg', 'image/jpg' or 'image/png'.",
        "operationId": "modifyOneUser",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "userId",
            "in": "path",
            "description": "Id of user to modify.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "User account properties modifications. For more details, go to 'UserPutJson' or 'UserPutFormdata' schemas. At least one of the 'user' object properties or an image is needed. 'isAdmin' property change will be taken into consideration only if it is made by an admin.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userPutJson"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/userPutFormdata"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "404": {
            "description": "Invalid Id."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete one account.",
        "description": "User deletes his own account or admin deletes a user account.",
        "operationId": "deleteOneUser",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "userId",
            "in": "path",
            "description": "Id of user to delete.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "404": {
            "description": "Invalid Id."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "userPostJson": {
        "title": "UserPostJson",
        "description": "User schema used when admin creates an account without any image.",
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string",
            "example": "Bob Marley"
          },
          "theme": {
            "type": "string",
            "example": "nuancesDeGris"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "monEmail@example.com"
          },
          "password": {
            "type": "string",
            "example": "MotDePasse"
          },
          "isAdmin": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "pseudo",
          "email",
          "password"
        ]
      },
      "userPostFormdata": {
        "title": "UserPostFormdata",
        "description": "User schema used when admin creates an account with an image. File type must be 'image/jpeg', 'image/jpg' or 'image/png'.",
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "pseudo": {
                "type": "string",
                "example": "Bob Marley"
              },
              "theme": {
                "type": "string",
                "example": "nuancesDeGris"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "monEmail@example.com"
              },
              "password": {
                "type": "string",
                "example": "MotDePasse"
              },
              "isAdmin": {
                "type": "boolean",
                "example": true
              }
            },
            "required": [
              "pseudo",
              "email",
              "password"
            ]
          },
          "image": {
            "type": "string",
            "example": "MonImage.jpg"
          }
        },
        "required": [
          "user",
          "image"
        ]
      },
      "userPutJson": {
        "title": "UserPutJson",
        "description": "User schema used when user changes an account but not the image. At least one property of 'user' object is required. 'isAdmin' property changes won't be taken into consideration if the user modifying is not an admin.",
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string",
            "example": "Bob Marley"
          },
          "theme": {
            "type": "string",
            "example": "nuancesDeGris"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "monEmail@example.com"
          },
          "password": {
            "type": "string",
            "example": "MotDePasse"
          },
          "isAdmin": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "bearerToken": {
        "description": "Bearer token schema.",
        "type": "string",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzMxZDk1MWU3N2YxNDAxMmQxMjExNmEiLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNjY0MjExMzEyLCJleHAiOjE2NjQyOTc3MTJ9.03h-xFcgHmwpYCHMZq36hnSAsHNS7TItboYmAliKSIk"
      },
...
    }
  }
}